third_party/WebKit/LayoutTests/http/tests/serviceworker/serviceworkerobject-scope.html

<!DOCTYPE html>
<title>ServiceWorker object: scope and url property</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>

function scope_test(name, url, scope) {
    var t = async_test('Verify the scope and url property: ' + name);

    t.step(function() {
        var expectedScope;
        if (scope)
            expectedScope = normalizeURL(scope);
        else
            expectedScope = normalizeURL('/*');
        var expectedURL = normalizeURL(url);

        service_worker_unregister_and_register(t, url, scope, onRegister);

        function onRegister(worker) {
            assert_equals(worker.scope, expectedScope, 'Returned ServiceWorker object should have scope');
            assert_equals(worker.url, expectedURL, 'Returned ServiceWorker object should have url');
            service_worker_unregister_and_done(t, scope);
        }
    });
}

scope_test('default', 'resources/worker-no-op.js');
scope_test('wildcard', 'resources/worker-no-op.js', '*');
scope_test('script with fragment', 'resources/worker-no-op.js#ref', 'http://127.0.0.1:8000/*');
scope_test('scope with fragment', 'resources/worker-no-op.js', 'http://127.0.0.1:8000/*#ref');
scope_test('scope with query', 'resources/worker-no-op.js', '/a?b=c');
scope_test('relative path', 'resources/worker-no-op.js', '/a/b/c/');
scope_test('relative path with wildcard', 'resources/empty-worker.js', '/a/b/c/*');
scope_test('absolute url', 'resources/empty-worker.js', 'http://127.0.0.1:8000/');
scope_test('absolute url with wildcard', 'resources/empty-worker.js', 'http://127.0.0.1:8000/*');

</script>
